+++ apps/web/src/middleware.ts
@@ -6,5 +6,7 @@ const { auth } = NextAuth(authConfig);
 export default auth;
 
 export const config = {
-	matcher: ["/((?!login|api/auth|_next/static|_next/image|favicon.ico).*)"],
+	matcher: [
+		"/((?!login|changelog|api/auth|_next/static|_next/image|favicon.ico).*)",
+	],
 };